# Refactor `SearchContainer` filtering operation and fix potential flaws