--- a/osu.Framework.Tests/Visual/UserInterface/TestSceneSearchContainer.cs
+++ b/osu.Framework.Tests/Visual/UserInterface/TestSceneSearchContainer.cs
@@ -144,7 +144,7 @@ namespace osu.Framework.Tests.Visual.UserInterface
             AddStep("Search term: " + term, () => textBox.Text = term);
         }
 
-        private class HeaderContainer : Container, IHasFilterableChildren
+        private class HeaderContainer : Container, IFilterable
         {
             public IEnumerable<string> FilterTerms => header.FilterTerms;
 
@@ -163,8 +163,6 @@ namespace osu.Framework.Tests.Visual.UserInterface
             {
                 set { }
             }
-
-            public IEnumerable<IFilterable> FilterableChildren => Children.OfType<IFilterable>();
 
             protected override Container<Drawable> Content => flowContainer;