--- a/osu.Framework/Graphics/Containers/FlowContainer.cs
+++ b/osu.Framework/Graphics/Containers/FlowContainer.cs
@@ -216,7 +216,7 @@ namespace osu.Framework.Graphics.Containers
 
             if (!childLayout.IsValid)
             {
-                layout.Invalidate();
+                InvalidateLayout();
                 childLayout.Validate();
             }