--- a/osu.Framework/Graphics/Containers/IFilterable.cs
+++ b/osu.Framework/Graphics/Containers/IFilterable.cs
@@ -3,15 +3,18 @@
 
 namespace osu.Framework.Graphics.Containers
 {
+    /// <summary>
+    /// Represents a component which can be filtered out on non-matching search terms.
+    /// </summary>
+    /// <remarks>
+    /// Hiding <see cref="IFilterable"/>s for purposes other than filtering require being wrapped in a
+    /// <see cref="VisibilityContainer"/> with its visibility state set to <see cref="Visibility.Hidden"/>.
+    /// </remarks>
     public interface IFilterable : IHasFilterTerms
     {
         /// <summary>
         /// Whether the current object is matching (ie. visible) given the current filter criteria of a parent.
         /// </summary>
-        /// <remarks>
-        /// It is recommended to not let this property update the <see cref="Drawable.Alpha"/> state of this filterable, as to not conflict with hiding it via other means.
-        /// But instead, override <see cref="Drawable.IsPresent"/> to respect the filter state, and invalidate <see cref="Invalidation.Presence"/> accordingly.
-        /// </remarks>
         bool MatchingFilter { set; }
 
         /// <summary>