@peppy 2022-05-05
Just a thought: Would making `MatchingFilter` a `get;set;` property and relying on that instead of `IsPresent` not clean things up? I'm still not fully on board with this PR due to the `IsPresent` dependency..

ie. for any containers osu-side which are hiding groups of settings, they would need to implement `IFilterable` and return `false` to `MatchingFilter` if they were hidden, or something like that..