@peppy 2022-05-05
Another alternative would be to *require* the use of `VisibilityContainer` for any "hiding" containers, and explicitly mention that in the `SearchContainer` usage instructions. Then we could rely on `VisibilityContainer.State` instead of `IsPresent` for the traversal checks.